+++ mora-core/src/entities/connections_info.rs
@@ -2,5 +2,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct ConnectionsInfo {
-    pub clients_connected: u64,
+    pub clients_connected: usize,
 }
